Delete is working

+++ static/script .js
@@ -6,7 +6,6 @@ function loadData(records = []) {
 	var table_data = "";
 	for(let i=0; i<records.length; i++) {
 		table_data += `<tr>`;
-
 		table_data += `<td>${records[i][0]}</td>`;
 		table_data += `<td>${records[i][1]}</td>`;
 		table_data += `<td>${records[i][2]}</td>`;
@@ -19,7 +18,7 @@ function loadData(records = []) {
 		table_data += `<td>`;
 		table_data += `<i class="fa fa-edit" style="font-size:28px;color:rgb(71, 105, 224)" onclick="redirectToUpdate(${records[i][0]})"></i>`;
 		table_data += '&nbsp;&nbsp;';
-		table_data += `<button class="btn btn-danger">Delete</button>`;
+		table_data += ` <i class="fa fa-trash-o" style="font-size:28px;color:red" onclick="deleteCollege(${records[i][0]},${i})"></i>`;
 		table_data += `</td>`;
 		table_data += `</tr>`;
 	}
@@ -54,7 +53,8 @@ function getCollegesById(id) {
 }
 
 function deleteData(id) {
-		fetch(`${api_url}?id=${id}`, {
+	if(confirm(`Are you sure you want to delete this record?`)){
+		fetch(`${api_url}/colleges?id=${id}`, {
 			method: "DELETE",
 			headers: {
 			  'Accept': 'application/json',
@@ -66,7 +66,11 @@ function deleteData(id) {
 		.then((data) => {
 			console.log(data); 
 		})
+		return true
+	}else{
+		return false
 	}
+}
